moved socialMediaIcons to atoms folder
